hoy gana river

+++ TP2(eda)/AllegroDraw.h
@@ -1,6 +1,7 @@
 #include <allegro5/allegro.h>
 #include <allegro5/allegro_primitives.h>
 #include <allegro5/allegro_font.h>
+#include <allegro5/allegro_ttf.h>
 
 #include "Backend.h"
 
@@ -11,7 +12,7 @@
 typedef struct {
 
 	ALLEGRO_FONT* font;
-	double* times_recorded;
+	long double* times_recorded;
 	int times_count;
 } FrontData;
